Adds the ability to set the starting content
This is useful in the hypothetical case users want to store some
configuration in the first log entry, as it has the useful property
of being as quick to access as the last log entry: O(1).

Given that hashing the contents is half the point of this, it is
strongly recommended that deterministic content not be placed in the
first log entry.